split generators into install and jasmine
