+++ lib/generators/lightning_cad/all_generator.rb
@@ -1,0 +1,10 @@
+require 'rails'
+
+module LightningCad
+  module Generators
+    class AllGenerator < ::Rails::Generators::Base
+      generate 'lightning_cad:install'
+      generate 'lightning_cad:jasmine'
+    end
+  end
+end
